# HeYuanERP_Server/specs/OpenAPI/openapi.yaml
openapi: 3.1.0
info:
  title: SmartSales2008 Modernized API
  version: 0.1.0
  summary: Plan A (.NET 8 + Vue) API template (parity with legacy)
  description: |
    Starter OpenAPI 3.1 spec for SmartSales2008 refactor. Replace placeholders
    and extend paths per PRD. Uses unified response envelope and pagination.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
  - url: http://localhost:5080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: keyword
          in: query
          schema: { type: string }
        - name: ownerId
          in: query
          schema: { type: string }
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedAccountList'
    post:
      tags: [Accounts]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Account'
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get account by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Account'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Accounts]
      summary: Update account
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Account'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    SizeParam:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Envelope:
      type: object
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data: {}
        traceId: { type: string, example: 4f1f9e9e-1a2b-3c4d-5e6f-1234567890ab }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            code: { type: string, example: ERR_VALIDATION }
            message: { type: string, example: invalid request }
            errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    Pagination:
      type: object
      properties:
        items:
          type: array
          items: {}
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
...
    Account:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        ownerId: { type: string }
        taxNo: { type: string }
        active: { type: boolean }
        lastEventDate: { type: string, format: date-time }
    AccountCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        ownerId: { type: string }
        taxNo: { type: string }
        description: { type: string }
    AccountUpdate:
      allOf:
        - $ref: '#/components/schemas/AccountCreate'
    PagedAccountList:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Account'
...